Add fixture support
Added fixture support. The fixture is an annotation, that is set on
class level. The fixture has one attribute, value, which is an Setup
class.

The setup class has a default implementation, which keeps track of
added clients and servers. The end user is suppose to implement the
initialize method, where clients and servers can be added.