Add assertion support
Added assertion support. The assertion support is built over AssertJ,
with the difference that the assertions are not triggered until
verify is called.

This allows the user to pre-define the assertions so that an expected
message can be created without having the actual message available.

Since AssertJ is now used in main, the scope is changed.