Better supplier support
The supplier support is upgraded, now each assert class does not have
to specify the assertThat(actual) each time.

Instead the new method getAssert() can be called.
This method is introduced in ObjectAssertImpl and can be overridden
in each assert class so that it returns the correct assert class.

This new way will reduce duplicate code since it will now be
possible to share code between classes, even if they doesn't
share any common interface, for example char[] and int[].

The EnumerableAssert now extends the ObjectAssert, so interfaces
extending this interface doesn't have to specify both EnumerableAssert
and ObjectAssert as interfaces.