Make assertions extendable
All the assertions should be extendable. It should be possible to be able
to define customized assertions and still have the chainging work.

With the current solution, that is not possible. This is because it's not
possible to change the return type on many of the assertions.

This is changed in this commit as all the assertions takes the return type
as a class parameter.

It's also introduced on interfaces with final classes, as String, Integer
and so on. Here, the type can't be changed, only the return type.