Remove Utils
In Java 12, it's no longer possible to change the value of final
fields.

Since that is changed, the utils class has no purpose and can be removed.

The DefaultExectutor is updated so it takes an argument in the contructor.
This is because we need to change this value from tests. The factory class
will always send in null, as was the default value earlier.