Add SetupTest
Added an integration test for the setup.
Since the setup is partly tested in ExecutorTest, some code could
be re-used. However, it was never a good idea to have the server static.

Instead, the server can be retrieved from the getServers method.
This will prevent multiple tests making changes to the same object.