Add reporter interface
Added the reporter interface. This interface makes it possible to
add reporters to Teacup. A reporter can handle different events, as
logs, started test cases etc.

All logs that should be sent to the different reporters, should use
the Factory class in the logging package. This will automatically
make sure that all reporters gets the logs.

The TestCase and TestSuite interfaces are used to represent test case
and test suite data.
The TestStatus class is used to list all the different statuses of a
test case.