Add test result interface
Since the reporter interface already can identify the different
stages in the test case life, we don't need to have a status for
each stage as well.

Therefor, we can remove the initialized, finished and skipped statuses.

The result is only provided when the test case has finished, since we need
to know if the test case passed, failed or was aborted for some reason.