Move test classes to testing package
Moved all classes related to test to its own package.

Also made changes to the hierarchy. A plan can contain data,
data can either be a case or a suite.

This makes it possible for the reporters to do tasks when for example
a test class is finished or started.
This was not possible earlier, since only test methods was reported.

There was also an potenital bug, if the test plan only contains a test,
without any suite. In that case, no test methods would be reported.

This solution makes the reporting more flexible and powerful.