Reduce duplicate code in testing package
The plan and suite was very similar in structure, and the
only thing differs between the plan and the suite is that the
suite is a defined container and the plan is just the "root".

Instead, plan is now called container, which the suite extends from.
The container, suite and case also extends from executable, which
handles the start/finish times.

The psysical interface holds the path to the psysical file, which can
either be a case or a suite.