Add tests containing tests and containers support
Some test engines might allow tests to contain containers, or other
tests.
Teacup should not put any restrictions on how the test engines wants to
execute tests. By allowing tests to contain tests and containers we now
have three layers.

Root, this is just used to mark the root of a test execution. In
JUnit it's called a TestPlan. It contains a start/stop time,
containers and executables.

Container, is a collection of containers and executables. The container
can, but doesn't have to contain other containers. The container,
can, but doesn't have to contain other executables.
It also contains the path to the container. This can be a test class in
JUnit.

Executable, has the same behavior as Container, with the addition that it contains a name. The
executable should be used for everything that can be executed as a test. This includes tests that
contains other tests, or containers. In JUnit, this can be a test method.