Add Node interface
The node can either be a container, a test or both.
The old interfaces container, executable and root are deleted together
with the impplementations.

The reason for this is because they all can do the same things.
The root needed a name, which it did not have earlier, the same
thing for the container.

The path is not needed since not everything is a path. It's better
to save the "path" in the name and let the reporter decide what it
want to do with the name/path.