Add more functionality to reporter
The reporter have two new methods: added and terminated.
The added method should handle event when nodes are added
after initialized is called.

The terminated should handle events when the test execution is
completed.

The method started(Iterable<Node>) is now called initialized, since
the naming is easy to confuse with the started(Node) method.