Add node to finished and log
The node is added to log method, because it's important that the method
knows that node log to write the logs to.

This could be tracked in each of the reporter implementations, but
that also means a lot of duplicate code.

Instead, this logic can be done in ReporterImpl, the class that Teacup
reports all events to and that handles the distribution of events
to all connected reporters.

Since we added this logic to the ReporterImpl, we can add Node to the
finished method as well, with the same argument as above.