Fix thread problem in log
When using Core's log factory, no node is passed into the
log method in ReporterImpl.

The problem was that the code assumed that only one thread was executing
the nodes. This is not correct and because of this, the logs where sometimes
written to the root log instead of the test log.

The solution to this is to traverse all active threads and look
for the ID to where the nodes where added from.